+++ project/software/display.h
@@ -8,7 +8,7 @@
 #ifdef SW
 #include <cv.h>
 
-typedef unsigned short * zbuffer_t;
+typedef unsigned int * zbuffer_t;
 
 typedef struct {
   int xres, yres;
@@ -16,7 +16,7 @@ typedef struct {
   zbuffer_t zbuffer;
 } gpImg;
 #else
-typedef volatile unsigned short * zbuffer_t;
+typedef volatile unsigned int * zbuffer_t;
 
 typedef struct {
   int xres, yres;
@@ -33,7 +33,7 @@ void gpSetImagePixel(gpImg *img, int x, int y, unsigned char r, unsigned char g,
 
 void gpSetImageHLine(gpImg *img, int y, int x1, int x2, unsigned char r, unsigned char g, unsigned char b);
 
-void gpSetImageHLineZBuff(gpImg *img, int y, int x1, int x2, unsigned short z1, unsigned short z2, unsigned char r, unsigned char g, unsigned char b);
+void gpSetImageHLineZBuff(gpImg *img, int y, int x1, int x2, unsigned int z1, unsigned int z2, unsigned char r, unsigned char g, unsigned char b);
 
 void gpDisplayImage(gpImg *img);
