+++ project/software/gp.h
@@ -16,8 +16,8 @@ typedef struct {
   int x, y;
 } gpVertex2Fixed;
 
-// typedef for z-buff unsigned shorts
-typedef unsigned short zbuff;
+// typedef for z-buff
+typedef unsigned int zbuff;
 
 // 3-d fixed point for rendering with zbuffer
 typedef struct {
